ref: move non-app files to warehouse/
