+++ examples/warehouse/app.py
@@ -3,9 +3,8 @@ import pandas as pd
 import solara
 from mesa.visualization import SolaraViz
 from mesa.visualization.utils import update_counter
-
-from .agents import InventoryAgent
-from .model import WarehouseModel
+from warehouse.agents import InventoryAgent
+from warehouse.model import WarehouseModel
 
 # Constants
 LOADING_DOCKS = [(0, 0, 0), (0, 2, 0), (0, 4, 0), (0, 6, 0), (0, 8, 0)]
